+++ CHANGELOG.md
@@ -10,6 +10,9 @@ A title is required for publishing a github release, so all versions should have
 
 ## Unreleased
 
+### Added
+- Add type aliases for `NonMax` and `NonZero` (i.e. `NonZeroU32`, `NonMaxU32`, etc.) (unqpkqks)
+
 ## 0.1.4 - 2026-08-25
 Make `NonZero::new`, `NonMax::{new, get}` a `const fn`.